feat(#475): PlannerService facade — logica verplaatst naar 5 use-case services
PlannerService.cs: 2118 → 50 regels (dunne facade)

Nieuwe services in FunctionApp/Planner/Services/:
  AvailabilityService   — CheckAvailabilityAsync + CheckDoordeweeksBeschikbaarAsync
  AutoPlanService       — AutoPlanAsync + AutoPlanToepassenAsync
  OptimizationService   — OptimaliseerAsync (incl. BerekenCapaciteit,
                          OptimaliseerGrasveldOntlasten, OptimaliseerStrakkerPlannen,
                          VerdeelExtraBuffer)
  RescheduleService     — CheckRescheduleAvailabilityAsync
  TeamScheduleService   — GetTeamScheduleAsync

PlannerShared           — gedeelde constanten, CanFitMatch, TryExactTime,
                          FindAllSlots, ToSlotToewijzing, AddWeekdayWarning,
                          RondAfOp5Min + CandidateSlot + IngeplandSlot + FieldScheduler

Alle 27 tests groen. Bestaande callers ongewijzigd.
Versie: 2.14.1.0 → 2.15.0.0 (feat = MINOR bump)
